Update README.md
Updated with URL to point to the beta of the newer version of the book we will be using. https://learnpythonthehardway.org/python3/

Kept links to Safari and Seattle Public Library if they want access to the older current edition and for when the newer edition becomes available so people could load the ePubs to their preferred readers/devices.